* 'boot-admin' of gitee.com:chundengtai/chundengtai_api:
  下单扣除积分

+++ cdt-infrastructure/cdt-base-bean/src/main/java/com/chundengtai/base/bean/CdtScoreFlow.java
@@ -44,7 +44,7 @@ public class CdtScoreFlow implements Serializable {
     /**
      * 积分总数
      */
-    private Long score;
+    private BigDecimal score;
 
     /**
      * 充值金额
